backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase==2.3.4

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Async Support
asyncio==3.4.3

# Environment Management
python-dotenv==1.0.0

# Data Validation
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.1.14
